+++ openghg/store/spec/__init__.py
@@ -3,4 +3,5 @@ from ._specification import (
     define_data_type_classes,
     define_standardise_parsers,
     define_transform_parsers,
+    check_parser,
 )